requirements.txt
aniso8601==3.0.2
argon2==0.1.10
argon2-cffi==18.3.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-configurations==2.1
# django-cors-headers==2.4.0
django-cors-headers==3.0.2
django-environ==0.4.5
django-eventtools==0.9.11
django-friendship==1.8.0
django-haystack==2.8.1
django-recurrence==1.9.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
elasticsearch>=5.0.0,<6.0.0
geographiclib==1.49
geopy==1.19.0
googlemaps==3.0.2
graphene==2.1.6
graphene-django==2.3.2
graphql-core==2.2
graphql-relay==0.4.5
idna==2.7
oauthlib==2.1.0
promise==2.2.1
psycopg2-binary==2.8.3
pycparser==2.19
PyJWT==1.6.4
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.23
